Split toolcall input parsing to satisfy line gate
